requirements.txt
Flask==1.0.2
psycopg2-binary==2.8.1
alembic==1.0.8
SQLAlchemy==1.3.2
Flask-GraphQL==2.0.0
flake8==3.7.7
graphene==2.1.6
flask-script==2.0.6
pytest==5.0.1
coverage==4.5.3
pyjwt==1.7.1
graphene-sqlalchemy==2.2.1
